fix-refactor/chat-feedback-1 : RedissonConfig 합치기 및 락 코드 간결화 피드백 반영
