--- src/main/java/com/dife/api/config/RedissonConfig.java
@@ -1,25 +1,0 @@
-package com.dife.api.config;
-
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RedissonConfig {
-
-	@Value("${spring.data.redis.host}")
-	private String redisHost;
-
-	@Value("${spring.data.redis.port}")
-	private int redisPort;
-
-	@Bean(destroyMethod = "shutdown")
-	public RedissonClient redisson() {
-		Config config = new Config();
-		config.useSingleServer().setAddress("redis://" + redisHost + ":" + redisPort);
-		return Redisson.create(config);
-	}
-}
